requirements.txt
# Core Dependencies
numpy>=1.24.0
torch>=2.0.0
sentence-transformers>=2.2.0
sounddevice>=0.4.6
soundfile>=0.12.1

# LLM Providers
anthropic>=0.25.0
openai>=1.0.0
ollama>=0.1.0

# TTS Providers
pyttsx3>=2.90
elevenlabs>=0.2.0

# STT Providers
faster-whisper>=0.10.0
vosk>=0.3.45
SpeechRecognition>=3.10.0
openai-whisper>=20231117

# Memory & Database
lancedb>=0.5.0
sqlite3  # Built-in

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Audio Processing
pyaudio>=0.2.11
sox>=1.4.1

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
pathlib2>=2.3.7
typing-extensions>=4.8.0 